+++ packages/core/src/api/factories/account/getAliases.ts
@@ -1,8 +1,10 @@
 /**
  * Copyright (c) 2019 Burst Apps Team
+ * Modified work Copyright (c) 2022 Signum Network
  */
 import {ChainService} from '../../../service/chainService';
 import {AliasList} from '../../../typings/aliasList';
+import {GetAliasesArgs} from '../../../typings/args/getAliasesArgs';
 
 /**
  * Use with [[ApiComposer]] and belongs to [[AccountApi]].
@@ -11,7 +13,10 @@ import {AliasList} from '../../../typings/aliasList';
  * @module core.api.factories
  */
 export const getAliases = (service: ChainService):
-    (accountId: string) => Promise<AliasList> =>
-    (accountId: string): Promise<AliasList> => service.query('getAliases', {
-        account: accountId,
+    (args: GetAliasesArgs) => Promise<AliasList> =>
+    (args: GetAliasesArgs): Promise<AliasList> => service.query('getAliases', {
+        account: args.accountId,
+        timestamp: args.timestamp,
+        firstIndex: args.firstIndex,
+        lastIndex: args.lastIndex
     });