+++ packages/core/src/typings/args/getAliasesArgs.ts
@@ -1,0 +1,25 @@
+/**
+ * The argument object for [[AccountApi.getAliases]]
+ *
+ * @module core
+ */
+export interface GetAliasesArgs {
+    /**
+     * The accountId id
+     */
+    accountId: string;
+
+    /**
+     * The timestamp (block time) you are looking for
+     */
+    timestamp?: number;
+
+    /**
+     * The first index of the list, beginning at 0
+     */
+    firstIndex?: number;
+    /**
+     * The last index of the list. At maximum 500 will be returned
+     */
+    lastIndex?: number;
+}